Create different file for every view component
