+++ DeUrgenta/Features/About/View/AboutView.swift
@@ -51,61 +51,3 @@ struct AboutView: View {
         }
     }
 }
-
-struct SpinnerView: View {
-    var body: some View {
-        HStack {
-            Spacer()
-
-            ZStack {
-                Text(AppStrings.AboutView.logoText.localized())
-                    .foregroundColor(.white)
-                    .customFont(.IBMPlexSansSemiBold, size: 24)
-                    .zIndex(1)
-
-                Image("spinner-large")
-                    .resizable()
-                    .aspectRatio(contentMode: .fit)
-                    .frame(width: 200, height: 200)
-            }
-
-            Spacer()
-        }
-    }
-}
-
-struct SocialMediaIconsView: View {
-    var body: some View {
-        HStack {
-            Image("facebook-icon")
-                .frame(width: 24, height: 24)
-                .padding()
-
-            Image("instagram-icon")
-                .frame(width: 24, height: 24)
-                .padding()
-
-            Image("link-icon")
-                .frame(width: 24, height: 24)
-                .padding()
-
-            Image("github-icon")
-                .frame(width: 24, height: 24)
-                .padding()
-        }
-    }
-}
-
-struct DonateButtonView: View {
-    var body: some View {
-        HStack {
-            Text(AppStrings.AboutView.donateButtonText.localized())
-                .foregroundColor(.secondary)
-                .customFont(.IBMPlexSansBold, size: 16)
-                .padding(.vertical)
-        }
-        .frame(maxWidth: 200, maxHeight: 50)
-        .background(Color.green)
-        .cornerRadius(6)
-    }
-}
